move types in authentication/signatures to typing.py
